moved source branch definition functions to branch_combinator.py
