+++ toshi_hazard_store/branch_combinator/branch_combinator.py
@@ -1,6 +1,8 @@
 # rom toshi_hazard_store.branch_combinator.SLT_test1 import *
 import itertools
 import math
+import json
+from collections import namedtuple
 
 DTOL = 1.0e-6
 
@@ -54,3 +56,58 @@ def get_weighted_branches(grouped_ltbs):
         raise Exception('weights do not sum to 1')
 
     return source_branches
+
+
+Member = namedtuple("Member", "group tag weight inv_id bg_id hazard_solution_id")
+
+
+def weight_and_ids(data):
+    def get_tag(args):
+        for arg in args:
+            if arg['k'] == "logic_tree_permutations":
+                return json.loads(arg['v'].replace("'", '"'))[0]['permute']  # ['members'][0]
+        assert 0
+
+    nodes = data['data']['node1']['children']['edges']
+    for obj in nodes:
+        tag = get_tag(obj['node']['child']['arguments'])
+        hazard_solution_id = obj['node']['child']['hazard_solution']['id']
+        yield Member(**tag[0]['members'][0], group=None, hazard_solution_id=hazard_solution_id)
+
+
+def all_members_dict(ltbs):
+    """LTBS from ther toshi GT - NB some may be failed jobs..."""
+    res = {}
+
+    def members():
+        for grp in ltbs[0][0]['permute']:
+            # print(grp['group'])
+            for m in grp['members']:
+                yield Member(**m, group=grp['group'], hazard_solution_id=None)
+
+    for m in members():
+        res[f'{m.inv_id}{m.bg_id}'] = m
+    return res
+
+
+def merge_ltbs(logic_tree_permutations, gtdata, omit):
+    members = all_members_dict(logic_tree_permutations)
+    # weights are the actual Hazard weight @ 1.0
+    for toshi_ltb in weight_and_ids(gtdata):
+        if toshi_ltb.hazard_solution_id in omit:
+            print(f'skipping {toshi_ltb}')
+            continue
+        d = toshi_ltb._asdict()
+        d['weight'] = members[f'{toshi_ltb.inv_id}{toshi_ltb.bg_id}'].weight
+        d['group'] = members[f'{toshi_ltb.inv_id}{toshi_ltb.bg_id}'].group
+        yield Member(**d)
+
+
+def grouped_ltbs(merged_ltbs):
+    grouped = {}
+    for ltb in merged_ltbs:
+        if ltb.group not in grouped:
+            grouped[ltb.group] = []
+        grouped[ltb.group].append(ltb)
+    return grouped
+
